requirements.txt
disputils
pandas
discord.py
pymongo